#PT1314922:93184114 Fixed parsing of pivotal tracker tag

+++ app/src/main/java/se/chalmers/eda397/pairprogramming/core/IPivotalTrackerClient.java
@@ -4,6 +4,6 @@ import se.chalmers.eda397.pairprogramming.model.UserStory;
 
 public interface IPivotalTrackerClient {
 
-    UserStory fetchUserStory(int projectId, int storyId);
+    UserStory fetchUserStory(long projectId, long storyId);
 
 }
